Added view for budget total

+++ TripGen/app/src/main/java/com/example/tripgen/Budget.java
@@ -60,7 +60,7 @@ public class Budget {
         categoryData.setBudget(budget);
     }
 
-    public double getBudget(Category category) {
+    public int getBudget(Category category) {
         CategoryData categoryData = getCategoryData(category);
         return categoryData.getBudget();
     }
